Helper was getting large, organize for convenience
